requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
arq==0.25.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.2.1
fastapi==0.109.2
fastcrud==0.12.1
gunicorn==22.0.0
httpx==0.26.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.6.1
python-jose==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
redis==5.2.0
sqlalchemy==2.0.36
sqlalchemy-utils==0.41.2
uvicorn==0.27.1
uvloop==0.19.0
email-validator==2.2.0
pytest==7.4.4
pytest-mock==3.14.0
Faker==26.3.0
google-generativeai
pillow
